+++ edit-gnu.c
@@ -36,7 +36,7 @@ static void edit_catcher(int sig) {
 
 static char *prompt;
 
-char *edit_alloc(void *cookie, int *count) {
+char *edit_alloc(void *cookie, size_t *count) {
 	struct cookie *c = cookie;
 	void (*oldint)(int), (*oldquit)(int);
